AVRO-3900: Merge the validators integration tests in one test
Just make sure the setup of the custom validators is done before the
parsing of any schema to prevent registering the default validator
(SpecificationValidator)
